refactor: :recycle: refactor homepage to single file
