+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:pushtotalk/pages/home_page.dart';
 
 void main() {
   runApp(const MyApp());
@@ -16,32 +17,7 @@ class MyApp extends StatelessWidget {
             seedColor: const Color.fromRGBO(97, 67, 209, 1)),
         useMaterial3: true,
       ),
-      home: const MyHomePage(),
-    );
-  }
-}
-
-class MyHomePage extends StatefulWidget {
-  const MyHomePage({super.key});
-
-  @override
-  State<MyHomePage> createState() => _MyHomePageState();
-}
-
-class _MyHomePageState extends State<MyHomePage> {
-  @override
-  Widget build(BuildContext context) {
-    return const Scaffold(
-      body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            Text(
-              'This is a text widget',
-            ),
-          ],
-        ),
-      ),
+      home: const HomePage(),
     );
   }
 }
